consensus: remove obsolete clique PoA engine
Clique proof-of-authority is no longer used by any supported network.
Remove the entire execution/protocol/rules/clique package and all
references: config structs, CLI flags, DB tables, RPC API registration,
and engine instantiation paths. Move SealHash/encodeSigHeader into the
aura package (the only remaining consumer). Simplify DeveloperGenesisBlock
to produce a non-clique genesis directly.
