+++ db/config3/config3.go
@@ -75,6 +75,6 @@ const (
 	// FullImmutabilityThreshold is the number of blocks after which a chain segment is
 	// considered immutable (i.e. soft finality). It is used by the downloader as a
 	// hard limit against deep ancestors, by the blockchain against deep reorgs, by
-	// the freezer as the cutoff threshold and by clique as the snapshot trust limit.
+	// the freezer as the cutoff threshold.
 	FullImmutabilityThreshold = 5_000
 )
