+++ execution/protocol/rules/merge/merge.go
@@ -66,7 +66,7 @@ var (
 //
 // Note: After the Merge the work is mostly done on the Consensus Layer, so nothing much is to be added on this side.
 type Merge struct {
-	eth1Engine rules.Engine // Original rules engine used in eth1, e.g. ethash or clique
+	eth1Engine rules.Engine // Original rules engine used in eth1, e.g. ethash
 
 	// Reusable buffer for collecting logs in Finalize - protected by logsBufMu
 	logsBufMu sync.Mutex
