+++ execution/tests/blockgen/chain_makers.go
@@ -95,8 +95,8 @@ func (b *BlockGen) SetNonce(nonce types.BlockNonce) {
 }
 
 // SetDifficulty sets the difficulty field of the generated block. This method is
-// useful for Clique tests where the difficulty does not depend on time. For the
-// ethash tests, please use OffsetTime, which implicitly recalculates the diff.
+// useful for tests where the difficulty does not depend on time. For the ethash
+// tests, please use OffsetTime, which implicitly recalculates the diff.
 func (b *BlockGen) SetDifficulty(diff uint64) {
 	b.header.Difficulty.SetUint64(diff)
 }
