+++ node/cli/default_flags.go
@@ -161,10 +161,6 @@ var DefaultFlags = []cli.Flag{
 	&utils.GpoPercentileFlag,
 	&utils.InsecureUnlockAllowedFlag,
 	&utils.IdentityFlag,
-	&utils.CliqueSnapshotCheckpointIntervalFlag,
-	&utils.CliqueSnapshotInmemorySnapshotsFlag,
-	&utils.CliqueSnapshotInmemorySignaturesFlag,
-	&utils.CliqueDataDirFlag,
 	&utils.ProposingDisableFlag,
 	&utils.MinerGasLimitFlag,
 	&utils.MinerEtherbaseFlag,
